fix(ensindexer): resolve circular dependency error
